requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pgvector==0.1.1
phidata==2.7.6
pillow==11.1.0
psycopg==3.1.10
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.31.0
rich==13.9.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.16
tomli==2.2.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
